move login components to login area component
